apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlmodel==0.0.14
pdfplumber==0.10.3
pymupdf==1.23.8
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
ruff==0.1.9
httpx==0.25.2
